feat(containers): Add deployment retry functionality
Add the ability to retry failed deployments without recreating them,
addressing scenarios where initial deployment messaging
to devices fails due to network issues or device unavailability.
